third_party/blink/web_tests/virtual/plz-service-worker/external/wpt/html/cross-origin-embedder-policy/cache-storage-reporting-service-worker.https-expected.txt

This is a testharness.js-based test.
FAIL COEP support on ServiceWorker. assert_equals: expected 1 but got 0
FAIL COEP-Report-Only support on ServiceWorker. assert_equals: expected 1 but got 0
Harness: the test ran to completion.
